Add focus management in dialogs
The code has been updated to better manage focus within dialogs. In DialogPane, modifications were done to direct the focus to the oldOwner if not null, rather than always requesting focus. In addition, the closeButton in the dialog and the ContentPane have been set to non-focus-traversable to prevent focus shifts to undesired nodes.