Refactor timeline usage in DrawerStackPane
Introduced a field to store the Timeline object in DrawerStackPane class which ensures that any currently running animation is stopped before starting a new one. This ensures smoother transitions between animations and reduces the likelihood of animations overlapping or interfering with each other. The "import" statement is also optimized to import javafx.animation.* instead of individual classes.