Refactor event handling and improve navigation in CalendarPicker
Reorganized mouse and touch events to occur in the CalendarPicker instead of CalendarPickerSkin. Altered key events to work with individual days instead of months for more precise navigation. Introduced a String Converter for parsing LocalDate types. The changes provide a more intuitive interaction with the calendar and streamline its functionality.