Add new ResponsivePane feature
This commit adds a new feature, ResponsivePane. This feature includes a pane that responds to layout changes based on conditions such as availability of space in the layout. The layout behavior when both small and large sidebars are visible can be customized. This commit also includes demo applications for the ResponsivePane and corresponding CSS files for styling.