Add PowerPane feature in gemsfx
Introduced a new PowerPane feature in the gemsfx library. This addition includes a requirement for com.dlsc.pdfviewfx in module-info.java and updates in the pom.xml file. PowerPane feature includes the combination of DialogPane, DrawerStackPane, InfoCenterPane, and a HiddenSidesPane all in one, making the UI more interactive and user-friendly. Also included demo implementation of the PowerPane in the demo package.