Relocate "showCloseButton" property and remove option from demo
Moved the "showCloseButton" property from DialogPane.java to the internal DialogHeader class. Consequently, removed the checkbox option to toggle "showCloseButton" property mainly from PowerPaneApp.java under the gemsfx-demo module and DialogsApp.java as it's no longer part of the public API. This enhances the encapsulation of the dialog display customization.