Update DialogPane and button visibility management
Modified the visibility management of DialogPane elements in the gemsfx module. Added an Observable listener to dynamically adjust the visibility and management of buttons according to the dialog type and presence of button types. Improved encapsulation by moving the 'showCloseButton' property internally. Also introduced a new Blank button to DialogsApp in the demo module for testing purposes.