Refactor `ResizablePane` into a non-visual `ResizingBehaviour` Class
The `ResizablePane` file has been converted into a non-visual `ResizingBehaviour` that can be attached to any region. This change aims to increase reusability, providing a resizing capability for any visual Node in the javafx scene, and to separate the resizing logic from the visual aspect of the Node. Additionally, the `ResizablePaneApp` file was renamed to `ResizingBehaviourApp`, reflecting the new change. An installation process was added as well. Finally, this refactoring was integrated into the `DialogPane` and `ContentPane` classes of the `gemsfx` package.