Adjust scene and stage sizing in EnhancedPasswordFieldApp
The scene and stage sizing has been updated in the EnhancedPasswordFieldApp. Specifically, the fixed size for scene has been removed to allow dynamic content sizing. Additionally, the stage is resized to fit the scene with the use of 'sizeToScene' method. This should improve application's flexibility in displaying different sizes of contents.