Refactored SearchTextField with additional features
The Refactoring of SearchTextField. Major changes include changes to allow the addition of text to the history when the user presses enter. Introduced methods to allow the addition or removal of history items. Added functions to display history popup to the user and its CSS aesthetics were also improved. Another notable change is the creation of a skin class resetting the selection of the ListView by focusing on it and selecting the first item. All these changes aim to provide a better user experience during search functionality.