Add option to store history on focus loss
A new option has been implemented in the SearchTextField class to allow the addition of text to history when the text field loses focus. The option is turned on by default but can be customized. The demo application has been updated to include a checkbox to trigger history addition on focus loss.