Add dynamic start angle functionality to CircleProgressIndicator
The update made to the CircleProgressIndicator and CircleProgressIndicatorSkin allows the start angle of the progress arc to be dynamically changed. This is done by binding a new startAngle property to the existing progressArc's startAngle property. The start angle of the progress arc can be customized using the setStartAngle method, which will update the visually displayed progress.