Refactor clickOnNode method and fix UIUtil comments
The clickOnNode method in the UIUtil class has been refactored to check if a mouse event is either a single or a multiple primary button click. An overloaded version of the method now accepts a boolean parameter to specify whether the event should be a single click or not. Additionally, minor fixes were done in the comments and format in UIUtil class. The method MouseClicked in EnhancedPasswordField.java was also modified to utilize the new clickOnNode method.